# openapi.yaml
...
openapi: 3.0.3
info:
  title: ScholarAI API
  description: AI-powered scholarship matching platform
  version: 5.0.0
  contact:
    name: ScholarAI Team
    email: support@scholarai.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.scholarai.com
    description: Production server
paths:
  /parse-cv:
    post:
      summary: Extract Profile from a PDF CV
      description: Upload a PDF CV and extract structured profile information using AI
      tags:
        - CV Processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cvFile:
                  type: string
                  format: binary
                  description: PDF file containing the CV (max 8MB)
              required:
                - cvFile
      responses:
        '200':
          description: Successfully extracted profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Bad request - missing file or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /pii-scrub/scrub:
    post:
      summary: Scrub PII from a profile object
      tags:
        - PII Scrubbing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  description: Profile object to scrub PII from
              required:
                - profile
      responses:
        '200':
          description: PII scrubbing results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIIScrubResult'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /pii-scrub/detect:
    post:
      summary: Detect PII in a profile object without scrubbing
      tags:
        - PII Scrubbing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: object
                  description: Profile object to analyze for PII
              required:
                - profile
      responses:
        '200':
          description: PII detection results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIIDetectionResult'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /pii-scrub/batch:
    post:
      summary: Scrub PII from multiple profiles
      tags:
        - PII Scrubbing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                profiles:
                  type: array
                  items:
                    type: object
                  maxItems: 50
                  description: Array of profile objects to scrub (max 50)
              required:
                - profiles
      responses:
        '200':
          description: Batch PII scrubbing results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPIIScrubResult'
        '400':
          description: Invalid input or batch size too large
        '500':
          description: Internal server error
  /pii-scrub/test:
    get:
      summary: Test PII scrubbing functionality
      tags:
        - PII Scrubbing
      responses:
        '200':
          description: Test results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  original:
                    type: object
                  scrubbed:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
                  error:
                    type: string
        '500':
          description: Internal server error
...
  /funding/classify:
    post:
      summary: Classify if a scholarship is fully funded
      tags:
        - Funding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scholarship
              properties:
                scholarship:
                  $ref: '#/components/schemas/Scholarship'
      responses:
        '200':
          description: Classification completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FundingClassification'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /funding/batch-classify:
    post:
      summary: Classify funding for multiple scholarships
      tags:
        - Funding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scholarships
              properties:
                scholarships:
                  type: array
                  items:
                    $ref: '#/components/schemas/Scholarship'
                  maxItems: 100
                concurrency:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 3
      responses:
        '200':
          description: Batch classification completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/BatchFundingClassification'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /funding/test:
    get:
      summary: Test the funding classification system
      tags:
        - Funding
      responses:
        '200':
          description: Test completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  testResults:
                    $ref: '#/components/schemas/BatchFundingClassification'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
  /funding/validate-classification:
    post:
      summary: Validate a funding classification result
      tags:
        - Funding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - classification
              properties:
                classification:
                  $ref: '#/components/schemas/ClassificationResult'
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: string
                  classification:
                    $ref: '#/components/schemas/ClassificationResult'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /funding/schema:
    get:
      summary: Get funding classification schema
      tags:
        - Funding
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deadline/parse:
    post:
      tags:
        - Deadline Parsing
      summary: Parse a single deadline text
      description: Parse deadline text into ISO YYYY-MM-DD format or "varies" using hybrid approach (AI with rule-based fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Deadline text to parse
                  example: "Application deadline: December 31, 2024"
      responses:
        '200':
          description: Deadline parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeadlineParseResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Profile:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: Full name of the applicant
        nationality:
          type: string
          nullable: true
          description: Nationality or citizenship
        degreeTarget:
          type: string
          enum: ["Bachelor", "Master", "PhD"]
          nullable: true
          description: Target degree level
        field:
          type: string
          nullable: true
          description: Field of study or specialization
        gpa:
          type: string
          nullable: true
          description: Grade point average in original format
        certifications:
          type: array
          items:
            type: string
          default: []
          description: List of certifications, test scores, licenses
          example: ["PMP", "IELTS 7.0", "AWS Certified"]
        specialStatus:
          type: string
          nullable: true
          description: Special status that may affect eligibility
          example: "refugee"
        languages:
          type: array
          items:
            type: string
          default: []
          description: Languages spoken by the applicant
          example: ["English", "Spanish", "French"]
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Additional error details (e.g., validation errors)
    PIIScrubResult:
      type: object
      properties:
        original:
          type: object
          description: Original profile object
        scrubbed:
          type: object
          description: Profile object with PII scrubbed
        piiDetected:
          type: array
          items:
            type: string
          description: List of PII types detected
        timestamp:
          type: string
          format: date-time
      required:
        - original
        - scrubbed
        - piiDetected
        - timestamp
    PIIDetectionResult:
      type: object
      properties:
        profile:
          type: object
          description: Original profile object
        piiDetected:
          type: array
          items:
            type: string
          description: List of PII types detected
        details:
          type: object
          description: Detailed PII detection information
        timestamp:
          type: string
          format: date-time
      required:
        - profile
        - piiDetected
        - timestamp
    BatchPIIScrubResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/PIIScrubResult'
        count:
          type: integer
          description: Number of profiles processed
        timestamp:
          type: string
          format: date-time
      required:
        - results
        - count
        - timestamp
    Scholarship:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the scholarship
        name:
          type: string
          description: Name of the scholarship
        country:
          type: string
          nullable: true
          description: Country where the scholarship is available
        degree:
          type: string
          enum: ["Bachelor", "Master", "PhD", "Any"]
          description: Degree level for the scholarship
        eligibility:
          type: string
          maxLength: 220
          description: Eligibility requirements (max 220 characters)
        deadline:
          type: string
          nullable: true
          pattern: '^(\\d{4}-\\d{2}-\\d{2}|varies)$'
          description: Application deadline in YYYY-MM-DD format or "varies"
        link:
          type: string
          format: uri
          description: Official application or information URL
        source:
          type: string
          description: Source domain of the scholarship information
        isFullyFunded:
          type: boolean
          description: Whether the scholarship is fully funded
      required:
        - id
        - name
        - degree
        - eligibility
        - link
        - source
        - isFullyFunded
    ClassificationResult:
      type: object
      required:
        - isFullyFunded
        - reason
      properties:
        isFullyFunded:
          type: boolean
          description: Whether the scholarship is fully funded
        reason:
          type: string
          maxLength: 160
          description: Explanation for the funding classification (max 160 characters)
    FundingClassification:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the classification was successful
        classification:
          $ref: '#/components/schemas/ClassificationResult'
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            model:
              type: string
              description: AI model used for classification
            scholarshipId:
              type: string
              description: ID of the classified scholarship
        error:
          type: string
          description: Error message if classification failed
    BatchFundingClassification:
      type: object
      properties:
        success:
          type: boolean
        results:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              scholarshipId:
                type: string
              success:
                type: boolean
              classification:
                $ref: '#/components/schemas/ClassificationResult'
              error:
                type: string
        summary:
          type: object
          properties:
            total:
              type: integer
            successful:
              type: integer
            failed:
              type: integer
            fullyFunded:
              type: integer
            partiallyFunded:
              type: integer
            processingTimeMs:
              type: integer
            averageTimePerScholarship:
              type: integer
        errors:
          type: array
          items:
            type: object
...
    DeadlineParseResult:
      type: object
      properties:
        success:
          type: boolean
          description: Whether parsing was successful
        deadline:
          type: string
          nullable: true
          pattern: '^(\d{4}-\d{2}-\d{2}|varies)$'
          description: Parsed deadline in YYYY-MM-DD format or "varies"
        originalText:
          type: string
          description: Original deadline text
        method:
          type: string
          enum: ["ai", "rules", "hybrid"]
          description: Method used for parsing
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score of the parsing result
        metadata:
          type: object
          properties:
            processingTimeMs:
              type: integer
            timestamp:
              type: string
              format: date-time
        error:
          type: string
          nullable: true
          description: Error message if parsing failed
      required:
        - success
        - originalText
        - method
        - confidence
        - metadata
...